+++ database/sql/src/main/scala/cromwell/database/sql/tables/CallCachingAggregationEntry.scala
@@ -5,5 +5,5 @@ case class CallCachingAggregationEntry
   baseAggregation: String,
   inputFilesAggregation: Option[String],
   callCachingEntryId: Option[Long] = None,
-  callCachingAggregationEntryId: Option[Int] = None
+  callCachingAggregationEntryId: Option[Long] = None
 )
